+++ src/v29rx.c
@@ -22,7 +22,7 @@
  * along with this program; if not, write to the Free Software
  * Foundation, Inc., 675 Mass Ave, Cambridge, MA 02139, USA.
  *
- * $Id: v29rx.c,v 1.99 2006/11/19 14:07:26 steveu Exp $
+ * $Id: v29rx.c,v 1.100 2006/11/28 16:59:57 steveu Exp $
  */
 
 /*! \file */
@@ -1546,8 +1546,8 @@ float v29_rx_symbol_timing_correction(v29_rx_state_t *s)
 void v29_rx_signal_cutoff(v29_rx_state_t *s, float cutoff)
 {
     /* The 0.4 factor allows for the gain of the DC blocker */
-    s->carrier_on_power = power_meter_level_dbm0(cutoff + 2.5f)*0.4f;
-    s->carrier_off_power = power_meter_level_dbm0(cutoff - 2.5f)*0.4f;
+    s->carrier_on_power = (int32_t) (power_meter_level_dbm0(cutoff + 2.5f)*0.4f);
+    s->carrier_off_power = (int32_t) (power_meter_level_dbm0(cutoff - 2.5f)*0.4f);
 }
 /*- End of function --------------------------------------------------------*/
